use UV as the package manager
One program could use multiple libraries, some are system package that comes with Python, some are 3rd party libs, like
pygame that you are using.

We need a way to track the libs we are using, and tell other people/developers about it.

There are multiple package management tools available, UV (https://docs.astral.sh/uv/) is the latest and so far the
best.

This commit also added a readme file about how to use it to get the program running